requirements.txt
playwright==1.40.0
python-dotenv==1.0.0
pytesseract==0.3.10
schedule==1.2.1
undetected-chromedriver==3.5.4
Pillow==10.1.0
